Use adapter for QueryReference requests
