+++ stream-chat-android-client/src/main/java/io/getstream/chat/android/client/ChatClient.kt
@@ -957,9 +957,6 @@ public class ChatClient internal constructor(
                 plugins.forEach { it.onQueryChannelsResult(result, request) }
             }
 
-    public fun queryChannelsNew(request: QueryChannelsRequest): QueryChannelsReference =
-        QueryChannelsReference(request = request, chatClient = this, scope = scope)
-
     @CheckResult
     public fun deleteChannel(channelType: String, channelId: String): Call<Channel> {
         return api.deleteChannel(channelType, channelId)
