+++ stream-chat-android-offline/src/main/java/io/getstream/chat/android/offline/plugin/QueryReference.kt
@@ -1,0 +1,9 @@
+package io.getstream.chat.android.offline.plugin
+
+import io.getstream.chat.android.client.call.Call
+import kotlinx.coroutines.CoroutineScope
+
+public interface QueryReference<T : Any, S : Any> {
+    public fun get(): Call<T>
+    public fun asState(scope: CoroutineScope): S
+}
